Refactor Slider to use mergeStyles
