+++ packages/office-ui-fabric-react/src/components/Slider/Slider.scss
@@ -1,134 +1,22 @@
 @import '../../common/common';
-
-.root {
-  user-select: none;
-}
-
-.titleLabel {
-  padding: 0;
- }
-
-.line {
-  display: flex;
-  position: relative;
-}
-
+// This shared between active and inactive spans. Need to create common and spread. Invert and use selectors: (?) https://github.com/OfficeDev/office-ui-fabric-react/wiki/Styling
 .line .lineContainer {
   border-radius: 4px;
   box-sizing: border-box;
 }
-
-.container {
-  display: flex;
-  flex-wrap: nowrap;
-  align-items: center;
-}
-
-.valueLabel {
-  flex-shrink: 1;
-  width: 30px;
-  line-height: 1;
-}
-
-.slideBox {
-  @include focus-border();
-  background: transparent;
-  border: none;
-  flex-grow: 1;
-  line-height: 28px;
-}
-
-.slideBox .thumb {
-  border: 2px solid $ms-color-neutralSecondary;
-  box-sizing: border-box;
-  background: $ms-color-white;
-  display: block;
-  width: 16px;
-  height: 16px;
-  position: absolute;
-  border-radius: 10px;
-}
-
+ 
+// Horizontal Variant
 .rootIsHorizontal {
-  .line {
-    width: 100%;
-  }
-
   .line .lineContainer {
     height: 4px;
   }
-
-  .slideBox {
-    height: 28px;
-    padding: 0 8px; // Make room for thumb at ends of line
-  }
-
-  .slideBox .thumb {
-    top: -6px;
-    @include transform-translateX(-50%);
-  }
-
-  .valueLabel {
-    margin: 0 8px;
-    white-space: nowrap;
-    width: 40px;
-  }
 }
 
+// Vertical Variant
 .rootIsVertical {
-  margin-right: 8px;
-
-  .line {
-    height: 100%;
-    width: 4px;
-    margin: 0 auto;
-    flex-direction: column-reverse;
-  }
-
   .line .lineContainer {
     width: 4px;
     height: 100%;
-  }
-
-  .slideBox {
-    height: 100%;
-    width: 28px;
-    padding: 8px 0; // Make room for thumb at bottom of line
-  }
-
-  .slideBox .thumb {
-    left: -6px;
-    margin: 0 auto;
-    transform: translateY(8px);
-  }
-
-  .container {
-    flex-direction: column;
-    height: 100%;
-    text-align: center;
-    margin: 8px 0;
-  }
-
-  .valueLabel {
-    margin: 0 auto;
-    white-space: nowrap;
-    width: 40px;
-  }
-}
-
-.activeSection {
-  background: $ms-color-neutralSecondary;
-
-  @include high-contrast {
-    background-color: WindowText;
-  }
-}
-
-.inactiveSection {
-  background: $ms-color-neutralTertiaryAlt;
-
-  @include high-contrast {
-    border: 1px solid WindowText;
   }
 }
 
@@ -136,43 +24,35 @@
   .thumb {
     transition: left $ms-animation-duration-3 $ms-animation-ease-1;
   }
-
   .activeSection,
   .inactiveSection {
     transition: width $ms-animation-duration-3 $ms-animation-ease-1;
   }
 }
 
+// Enabled Variant
 .rootIsEnabled {
-
   .slideBox:hover,
   .slideBox:active {
-
     .thumb {
       border: 2px solid $ms-color-themePrimary;
-
       @include high-contrast {
         border-color: Highlight;
       }
     }
-
     .activeSection {
       background-color: $ms-color-themePrimary;
-
       @include high-contrast {
         background-color: Highlight;
       }
     }
-
     .inactiveSection {
       background-color: $ms-color-themeLight;
-
       @include high-contrast {
         border-color: Highlight;
       }
     }
   }
-
   .slideBox:active {
     .thumb {
       border: 2px solid $ms-color-themeDarkAlt;
@@ -183,15 +63,14 @@
   }
 }
 
+// Disabled Variant
 .rootIsDisabled {
   .thumb {
     border-color: $ms-color-neutralTertiaryAlt;
-
     @include high-contrast {
       border-color: GrayText;
     }
   }
-
   .activeSection,
   .inactiveSection {
     @include high-contrast {
@@ -199,7 +78,6 @@
       border-color: GrayText;
     }
   }
-
   .activeSection {
     background: $ms-color-neutralTertiaryAlt;
   }
