+++ packages/office-ui-fabric-react/src/components/Slider/index.ts
@@ -1,2 +1,3 @@
 export * from './Slider';
+export * from './Slider.base';
 export * from './Slider.types';
